requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.8.0
async-timeout==4.0.3
attrs==23.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.6.0
azure-functions==1.17.0
azure-functions-durable==1.2.9
azure-identity==1.16.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-core==1.4.0
azure-monitor-query==1.4.0
azure-storage-blob==12.19.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorclass==2.2.2
colour==0.1.5
compressed_rtf==1.0.6
cryptography==44.0.0
distro==1.9.0
easygui==0.98.3
ebcdic==1.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
extract-msg==0.52.0
frozenlist==1.4.1
furl==2.1.3
h11==0.14.0
httpcore==1.0.4
httpx==0.28.1
idna==3.6
importlib_metadata==8.5.0
isodate==0.6.1
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.1.8
lxml==5.1.0
Markdown==3.7
msal==1.28.0
msal-extensions==1.1.0
msoffcrypto-tool==5.4.1
msrest==0.7.1
multidict==6.0.5
numpy==1.26.3
oauthlib==3.2.2
olefile==0.47
oletools==0.60.1
openai==1.99.2
openpyxl==3.1.5
orderedmultidict==1.0.1
packaging==24.0
pandas==2.2.2
pcodedmp==1.2.6
pillow==11.0.0
portalocker==2.8.2
psycopg2-binary==2.9.9
pyarrow==15.0.0
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymssql==2.3.7
PyMuPDF==1.25.5
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==1.1.2
python-pptx==0.6.23
pytz==2023.3.post1
red-black-tree-mod==1.20
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.22.1
RTFDE==0.1.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
styleframe==4.2
tabulate==0.9.0
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
websocket-client==1.7.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.21.0
git+https://github.com/estyle-llm/i-style.git@1.24.0